requirements.txt
aiohttp==3.7.3
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
idna==3.1
multidict==5.1.0
numpy==1.19.5
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
